requirements.txt
flask==2.3.3
gunicorn==21.2.0
faiss-cpu
openai>=1.0.0
python-dotenv==1.0.1
transformers>=4.40.2
sentence-transformers==2.6.1
torch>=1.13.0
numpy